Game can be started on pre-defined maps (#168)
In folder `data/maps`, maps in ascii format can be provided for different starting landscapes. Map symbols are provided in `data/json/terrain.json`.

Each map can have its own probabilities for random terrain cards, to allow for continued play in the map's spirit.

Further, the main menu was split up into tabs. One for starting a new game (from a map), and one for loading a game. This gives to opportunity to easily add a Settings tab, e.g. for the tileset or for an enlarged UI (see #166).

# Commits 

* get map symbols from JSON
* translate map to string
* save map to ascii file
* prepare loading maps
* add a system to load a map
* move load panel to separate function
* move new world stuff to separate container
* move new world and load world to separate tabs
* list available maps
* add a label for maps
* prepare for embedded maps
* add starting tile position to map ascii
* move scenarios to a separate tab
* add swamp forest map
* add desert valley map
* update changelog
* random terrains are stored in maps
* fix listing games and maps in WASM build